Revert "fixup! Rename client_callback.h"
This reverts commit e01937ec2268d03fa578b113fa74da5b5f42da3c.
